pyproject.toml
[project]
name = "cot-oracle"
version = "0.1.0"
description = "Detect unfaithful chain-of-thought reasoning via activation trajectory oracles"
requires-python = ">=3.10"
dependencies = [
    # Core ML — pin to match requirements.txt
    # torch 2.7+ has gradient checkpointing bug with forward hooks (tensor count mismatch)
    "torch>=2.6.0,<2.7.0",
    "transformers==4.55.2",
    "peft==0.17.1",
    "bitsandbytes>=0.48.1",
    "accelerate>=1.10.1",
    "datasets>=3.6.0",
    "numpy<2.0.0",
    "pandas",
    "tqdm",
    "scikit-learn",
    "scipy",
    "matplotlib",
    "seaborn",
    "huggingface_hub",
    "wandb>=0.21.1",
    "jiter",
    "httpx",
    "anthropic",
    "openai",
    "slist",
    "tenacity",
    "python-dotenv",
    "einops>=0.8.1",
    "jupyterlab",
    "ipywidgets",
    "vllm",
    "flask",
]

[tool.cot-oracle]
system-requirements = ["cloudflared"]
